refactor/moved logic to hooks
